mycode/requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
bleach==2.1.2
blessings==1.6
blinker==1.3
bpython==0.15
Brlapi==0.6.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
curtsies==0.2.6
decorator==4.1.2
defer==1.0.6
defusedxml==0.5.0
dirspec==13.10
Django==1.11.5
django-hosts==3.0
django-tinymce4-lite==1.7.0
entrypoints==0.2.3
eulerlib==0.2
flake8==3.5.0
greenlet==0.4.9
html5lib==1.0.1
httplib2==0.9.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.3
keras-tqdm==2.0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.14.0
oauthlib==2.0.6
onboard==1.2.0
oneconf==0.3.9
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==3.1.2
piston-mini-client==0.7.5
progressbar2==3.34.3
prompt-toolkit==1.0.15
psutil==3.4.2
ptyprocess==0.5.2
pyasn1==0.1.9
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.5.3
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-utils==2.2.0
python-xapp==1.0.0
python3-openid==3.1.0
pytz==2017.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
reportlab==3.3.0
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.0
sessioninstaller==0.0.0
setproctitle==1.1.8
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
system-service==0.3
terminado==0.8.1
testpath==0.3.1
thin-client-config-agent==0.8
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.22
virtkey==0.63.0
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
xkit==0.0.0
yapf==0.20.1
